Catch nan/inf arguments to buffering sooner. Report exceptions in geosop instead of dying
